==> Nonessential_MATa_screen/src/01_rawDataQualityControl.Rmd <==
---
title: "Raw data quality control"
author: "Ashwini Kumar Sharma, PhD"
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    highlight: kate
    code_folding: hide
---

```{r knitr_init, echo=FALSE, cache=FALSE}
library(knitr)
library(rmdformats)

## Global options
options(max.print=100000)
opts_chunk$set(echo=TRUE,
	             cache=FALSE,
               prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=80)
```

***

## Initialization

We start the analysis by initializing the packages required for all the analysis performed in this section. We also define the root directory, within which all the input/output operations for this project will be performed. At the [end of this document](#SI), detailed software version information is provided for easier reproducibility of the analysis.

```{r}
library(reshape2)
library(ggplot2)
library(ggrepel)
library(ggpubr)
library(lsr)
library(plotly)
library(DT)
library(WriteXLS)
library(ggallin)
path = "/Users/ashwin/Documents/Projects/YeastScreen/Nonessential_MATa_screen/"
```

***

## Raw data summary

We show below the summary for the raw screening data - Yeast Non-essential MATa library.

```{r}
rawDat = readRDS(paste0(path,"data/workspaces/YeastRedox_MATa_RawData.RDS"))
rawDat$Signal_485_520 = log10(rawDat$Signal_485_520)
rawDat$Signal_400_520 = log10(rawDat$Signal_400_520)

summary(rawDat)
```

## Background Fluorescence

Firstly, we look at the distribution of background flourescence across all plates and retain only those colonies for Control, Mitochondria and Cytoplasm whose fluorescence signal exceeds the 90th percentile of blank 485-520 and 400-520 fluorescence signals

```{r, figure.height = 8}

tmpdat = rawDat[,c("Plate", "Content", "Signal_485_520", "Signal_400_520", "Type")]
tmpdat = melt(tmpdat)

a = droplevels(tmpdat[tmpdat$variable == "Signal_485_520" & tmpdat$Content == "Blank",])
per90_485_520 = lapply(split(a, a$Plate), function(x) split(x, x$Type))
per90_485_520  = sapply(per90_485_520 , function(x) sapply(x, function(y) quantile(y$value, 0.90)))
per90_485_520  = melt(t(per90_485_520 ))
per90_485_520$Var2 = gsub(".90%", "", per90_485_520 $Var2)
colnames(per90_485_520) = c("Plate", "Type", "value")

a = droplevels(tmpdat[tmpdat$variable == "Signal_400_520" & tmpdat$Content == "Blank",])
per90_400_520 = lapply(split(a, a$Plate), function(x) split(x, x$Type))
per90_400_520 = sapply(per90_400_520 , function(x) sapply(x, function(y) quantile(y$value, 0.90)))
per90_400_520 = melt(t(per90_400_520 ))
per90_400_520$Var2 = gsub(".90%", "", per90_400_520$Var2)
colnames(per90_400_520) = c("Plate", "Type", "value")

# Filtering out colonies

if(identical(paste0(per90_400_520$Plate, per90_400_520$Type),
             paste0(per90_485_520$Plate, per90_485_520$Type))){
  cutoff = data.frame(per90_400_520[,1:2], 
                      per90_400_520 = per90_400_520$value,
                      per90_485_520 = per90_485_520$value,
                      stringsAsFactors = F)
}

rawDatFilt = lapply(split(rawDat, rawDat$Plate), function(x) split(x, x$Type))

for(i in 1: nrow(cutoff))
{
  x = rawDatFilt[[ as.character(cutoff$Plate[i]) ]][[cutoff$Type[i]]]
  x = x[x$Signal_485_520 > cutoff$per90_485_520[i] & x$Signal_400_520 > cutoff$per90_400_520[i],]
  rawDatFilt[[ as.character(cutoff$Plate[i]) ]][[cutoff$Type[i]]] = x
}

rawDatFilt = do.call("rbind", lapply(rawDatFilt, function(x) do.call("rbind", x)))

# Plot background fluorescence distributions
p1 = ggplot(tmpdat, aes(x = value, fill = Content)) + 
    geom_density(alpha =0.5) + theme_bw(base_size = 8) + 
    labs(subtitle = "Non-essential MATa screen", 
         x = "Fluorescence  signal (in log10)", 
         y = "Density", fill = "") +
    facet_wrap(variable ~ Type, scales = "free", ncol = 3) +
    theme(legend.position = "right", panel.grid = element_blank())

p2 = ggplot(tmpdat[tmpdat$variable == "Signal_485_520",], 
            aes(x = value, fill = Content)) + 
            geom_density(alpha =0.5) + theme_bw(base_size = 8) + 
            labs(subtitle = "Non-essential MATa screen | Signal_485_520", 
                 x = "Fluorescence  signal (in log10)", 
                 y = "Density", fill = "") +
            geom_vline(data = per90_485_520, aes(xintercept = value)) +
            facet_wrap(Plate ~ Type, scales = "free", ncol = 3) +
            theme(legend.position = "right", panel.grid = element_blank())

p3 = ggplot(tmpdat[tmpdat$variable == "Signal_400_520",],
            aes(x = value, fill = Content)) + 
            geom_density(alpha = 0.5) + theme_bw(base_size = 8) + 
            labs(subtitle = "Non-essential MATa screen | Signal_400_520", 
                 x = "Fluorescence  signal (in log10)", 
                 y = "Density", fill = "") +
            geom_vline(data = per90_400_520, aes(xintercept = value)) +
            facet_wrap(Plate ~ Type, scales = "free", ncol = 3) +
            theme(legend.position = "right", panel.grid = element_blank())

ggsave(filename = paste0(path,"analysis/QC/background_fluorescence.pdf"), plot = p1, width = 7, height = 3.5)
ggsave(filename = paste0(path,"analysis/QC/background_fluorescence_Signal_485_520.pdf"), plot = p2, width = 7, height = 60, limitsize = F)
ggsave(filename = paste0(path,"analysis/QC/background_fluorescence_Signal_400_520.pdf"), plot = p3, width = 7, height = 60, limitsize = F)

p1
```


```{r}
paste0("total number of colonies in the raw data - ", nrow(rawDat))
paste0("total number of colonies in the background flourescence filtered data - ", nrow(rawDatFilt))
paste0("% of colonies KEPT - ", round(c(nrow(rawDatFilt)/ nrow(rawDat) * 100), 2))

rawDat = rawDatFilt
```

```{r, echo=F}
rm(tmpdat, p1, p2, p3, per90_400_520, per90_485_520, a, i, x, cutoff, rawDatFilt)
```

## Cleaning raw data

Firstly we read the raw yeast redox screen data and remove outlier (extreme values) and `NA`. Typically, outlier removal is not considered good data analysis practice, however, we do it here because we know that the extreme values comes from technical artifacts. For example, roGFP2 ratios cannot be negative and those which are significantly higher than the plate average are ambiguous signals.

Outliers were detected as -

* upper outlier values > 75th quantile value for a plate + 3 * plate IQR value
* lower outlier values < 25th quantile value for a plate - 3 * plate IQR value

The outlier values are all set to ```NA``` and values greater than the lower bound threshold but less than 0 were set to pseudo minimum value of 0.0001 as roGFP2 ratios can't be less than 0.

We show these extreme outliers below, ploting the distribution of values from each plate.

```{r, fig.width = 10}
tmpdat = na.omit(rawDat)

ggplot(tmpdat, aes(x = roGFP2.ratio, color = Plate)) + 
  geom_density() + theme_bw(base_size = 8) + labs(x = "roGFP2 ratio (in log10)") +
   scale_x_continuous(trans = pseudolog10_trans, breaks = c(-50, -2:5, 50)) + facet_wrap(Content ~ Type, scales = "free") +
  theme(legend.position = "none", panel.grid = element_blank())

```

```{r, echo=F}
rm(tmpdat)
```


```{r}
rawDatCleaned = vector("list", length = 3)
names(rawDatCleaned) = c("Glucose", "Galactose", "Glycerol")

for(i in names(rawDatCleaned))
{
  dat = rawDat[rawDat$Type == i,]
  dat = split(dat, dat$Plate)
  dat = lapply(dat, function(x){
                                 vals = x$roGFP2.ratio
                                 lb = quantile(vals, probs = 0.25, na.rm=T) - (3 * IQR(vals, na.rm=T))
                                 ub = quantile(vals, probs = 0.75, na.rm=T) + (3 * IQR(vals, na.rm=T))
                                  
                                 vals[vals < lb | vals > ub] = NA
                                 vals[vals > lb & vals < 0] = 10^-4
                                  
                                 x$roGFP2.ratio = vals
                                 x = x[!is.na(x$roGFP2.ratio),]
                                 
                                 return(x)})
  
  rawDatCleaned[[i]] = do.call("rbind", dat)
  rm(dat)
}

rawDatCleaned = do.call("rbind", rawDatCleaned)
rawDatCleaned = droplevels(rawDatCleaned)
rownames(rawDatCleaned) = 1:nrow(rawDatCleaned)
```

Next, lets check the summaries of the raw roGFP2 redox screen data and the cleaned raw data for comparison.

Raw data looked like -

```{r}
summary(rawDat)
```

Cleaned data looked like -

```{r}
summary(rawDatCleaned)
```

Percentage of data removed after outlier removal -

```{r}
round((c(nrow(rawDat) - nrow(rawDatCleaned))/nrow(rawDat)) * 100,2)
```

```{r, echo=F}
rm(i, rawDat)
```

***

## Distribution of raw roGFP2 ratios

Distribution of raw roGFP2 ratios (absolute ratios) per plate

```{r, fig.width = 20, fig.height = 8}
p = ggplot(rawDatCleaned) + theme_bw(base_size = 8) +
geom_boxplot(aes(x = Plate, y = roGFP2.ratio), outlier.size = 0.1, lwd=0.2) +
facet_grid(Content ~ Type, scales = "free") +
theme(axis.text.x = element_text(angle = 90, hjust = 0.5, vjust = 0.5))

ggsave(filename = paste0(path,"analysis/QC/roGFP2_ratio_RAWdata_distribution_nutrient_compartments.pdf"), plot = p, width = 12, height = 5)

p

```

***

## Identification of edge effects

A well known technical issue with yeast mutant screens is the altered growth characteristics of yeast colonies growing at the edge of plates. Next, we look at how roGFP2 ratios are different between edge wells and non-edge wells. 

These are the edge wells and the mutants originating from these locations should be carefully analyzed.

```{r}
edge = as.character(unlist(read.csv(paste0(path,"data/annotations/edge_wells.txt"))))

# Print edge wells
edge

edgeDat = rep("non edge wells", nrow(rawDatCleaned))
edgeDat[rawDatCleaned$Well %in% edge] = "edge wells"

edge = rawDatCleaned
edge$edge = edgeDat; rm(edgeDat)
```

Visulaizing the differences using both effect size (Fold change) and pvalue. 

> For none of the plates and conditions, the edge effects exceed the log fold change of +/- 1

```{r}
# Computing distribution difference statistics
stats = split(edge, paste(edge$Plate, edge$Type, edge$Content, sep="-"))
stats = t(sapply(stats, function(x){
  
  y = x$roGFP2.ratio[x$edge == "edge wells"]
  n = x$roGFP2.ratio[x$edge== "non edge wells"]
  
  fc = log2(median(y, na.rm=T)/median(n, na.rm=T))
  pv = wilcox.test(y, n)$p.value
  #cd = effectsize::cohens_d(y, n, pooled_sd = F)$Cohens_d
  #p = signif(p,2)
  #return(c(log2fc = fc, wilcoxP = pv, CohenD = cd))
  return(c(log2fc = fc, wilcoxP = pv))
}))

a = do.call("rbind", strsplit(rownames(stats), "-"))
colnames(a) = c("Plate", "Nutrient", "Condition")
stats = data.frame(a, stats, row.names = NULL, stringsAsFactors = F)

stats$log2fc = round(stats$log2fc, 2)
stats$wilcoxP = signif(-log10(p.adjust(stats$wilcoxP, method="fdr")),2)
#stats$CohenD = round(stats$CohenD, 2)

# Volcano plot
p3 = ggplot(stats, aes(x = log2fc, y = wilcoxP, text = Plate)) + 
theme_bw(base_size = 8) + xlim(-1,1) + labs(y = "-log10(fdr pval)") +
geom_point(size = 0.8) + geom_hline(yintercept = -log10(0.05)) +
#geom_vline(xintercept = c(-0.5, 0.5), lty = 2) +
geom_vline(xintercept = c(-1, 1), lty = 1) +
facet_wrap(Nutrient ~ Condition, scales = "free") +
geom_text_repel(data = stats[abs(stats$log2fc) > 0.5 & stats$wilcoxP > -log10(0.05),], 
                aes(label = Plate), size = 2) + theme(panel.grid = element_blank())

ggsave(filename = paste0(path,"analysis/QC/roGFP2_ratio_RAWdata_edge_effects_badPlates.pdf"), 
       plot = p3, width = 7.5, height = 6)

ggplotly(p3)
```


```{r, echo =FALSE}
rm(edge, a)
```

***

## Variance among controls

Next we look into the variance exhibited by the control colonies across plates and nutrient conditions. Ideally, all control replicates for a given mutant in a given plate and nutrient condition should have similar values i.e variance close to zero. Below we show the variance of raw roGFP2 ratios among controls per plate and group.

```{r}
ctrlVar = rawDatCleaned %>% filter(Content == "Control") %>% select(Gene.Symbol, 
    Plate, Type, Group, roGFP2.ratio) %>% group_by(Plate, Type, Group) %>% mutate(Control_variance = round(var(roGFP2.ratio, 
    na.rm = T), 4)) %>% ungroup() %>% select(Gene.Symbol, Plate, Type, Group, Control_variance) %>% 
    rename(Genes = Gene.Symbol, Nutrient = Type) %>% mutate(Plate = paste0("P-", 
    Plate)) %>% distinct()

plot_ly(ctrlVar, x = ~Group, y = ~Plate, z = ~Control_variance, text = ~Genes, type = "scatter3d", 
    mode = "markers", marker = list(size = 3), opacity = 1, color = ~Nutrient)
```

Next we remove all genes with variance among controls > 0.05

```{r}
rmv = ctrlVar[which(ctrlVar$Control_variance > 0.05),]
rmv$Plate = gsub("P-", "", rmv$Plate)
rmv = paste(rmv$Genes,rmv$Plate, rmv$Nutrient, sep="-")

paste0("Number of poor control genes removed - ", length(rmv))

matchID = paste(rawDatCleaned$Gene.Symbol,rawDatCleaned$Plate, rawDatCleaned$Type, sep="-")
rawDatCleaned_ctrlfilt = rawDatCleaned[! matchID %in% rmv,]

ctrlVar = rawDatCleaned_ctrlfilt %>% filter(Content == "Control") %>% select(Gene.Symbol, 
    Plate, Type, Group, roGFP2.ratio) %>% group_by(Plate, Type, Group) %>% mutate(Control_variance = round(var(roGFP2.ratio, 
    na.rm = T), 4)) %>% ungroup() %>% select(Gene.Symbol, Plate, Type, Group, Control_variance) %>% 
    rename(Genes = Gene.Symbol, Nutrient = Type) %>% mutate(Plate = paste0("P-", 
    Plate)) %>% distinct()

plot_ly(ctrlVar, x = ~Group, y = ~Plate, z = ~Control_variance, text = ~Genes, type = "scatter3d", 
    mode = "markers", marker = list(size = 3), opacity = 1, color = ~Nutrient)

```

```{r}
rm(rmv, matchID, ctrlVar)
```

Next we we visualize the fact that in a good screen, individual control values should not deviate much from the median value of all controls from that same plate. For a given plate $i$, the normalized control values i.e $NormCtrl$ is given as - $$NormCtrl_i = \frac{Ctrl_i} {median(Ctrl_i)}$$ We plot these scaled control values per plate across conditions below. Ideally these values should be close to 1 (shown as red horizontal lines). A table for these scaled control values is also provided.

```{r, message=FALSE, fig.width=15}
ctrlDat = split(rawDatCleaned_ctrlfilt, rawDatCleaned_ctrlfilt$Type)
ctrlDat = lapply(ctrlDat, function(x){split(x, x$Plate)})

for( i in 1: length(ctrlDat))
{
  for(j in 1: length(ctrlDat[[i]]))
  {
    tmp = droplevels(ctrlDat[[i]][[j]])
    tmp = tmp[, -2]
  
    #Plate control
    tmp.ctrl = tmp[tmp$Content == "Control",]

    #Normalizing factor
    nf = median(tmp.ctrl$roGFP2.ratio, na.rm=T)
    
    # Measure control value deviation from plate median
    ctrlDat[[i]][[j]] = data.frame(Gene = tmp.ctrl$Gene.Symbol, 
                                   NormControl = round(log2(tmp.ctrl$roGFP2.ratio/nf), 3),
                                   stringsAsFactors = F)
    #Deleting
    rm(tmp, tmp.ctrl,  nf)
  }
  rm(j)
 }
rm(i)

ctrlDat = melt(ctrlDat, id = "Gene")
ctrlDat = ctrlDat[,c(4,5,1,3)]
colnames(ctrlDat) = c("Plate", "Nutrient", "Gene", "NormControl")
ctrlDat$Plate = as.factor(ctrlDat$Plate)
ctrlDat$Nutrient = factor(ctrlDat$Nutrient, levels = c("Glucose", "Galactose", "Glycerol"))

# datatable(ctrlDat, rownames = FALSE, filter="top", class="compact",
#           extensions = c('Buttons') , 
#           options = list(autoWidth = TRUE,
#                           dom = 'Blfrtip', 
#                           buttons = c('csv', 'excel')
#                         ))

p = ggplot(ctrlDat, aes(x = Plate, y = NormControl)) +
    theme_bw(base_size = 8) + labs(y="Control / Median plate control (roGFP2 ratios) in log2 scale") +
    geom_boxplot(outlier.size = 0.1, lwd=0.2) +
    facet_wrap( ~ Nutrient) +
    geom_hline(yintercept = 3, color = "orange") +
    geom_hline(yintercept = -3, color = "orange") +
    #geom_text_repel(data = subset(ctrlDat, NormControl > 3), aes(x = Plate, y = NormControl, label = Gene), size = 2) +
    theme(axis.text.x = element_text(angle = 90, hjust = 0.5, vjust = 0.5)) 

p
```

Next we remove those controls that deviate 3 log fold units over the plate control median value

```{r}
rmv = ctrlDat[which(ctrlDat$NormControl > 3 | ctrlDat$NormControl < -3),]
rmv = paste(rmv$Gene,rmv$Plate, rmv$Nutrient, sep="-")

ctrlDat = ctrlDat[!(ctrlDat$NormControl > 3 | ctrlDat$NormControl < -3),]

paste0("Number of control genes removed which deviated 3 log fold units from the plate median - ", length(rmv))

matchID = paste(rawDatCleaned_ctrlfilt$Gene.Symbol, 
                rawDatCleaned_ctrlfilt$Plate, 
                rawDatCleaned_ctrlfilt$Type, sep="-")
rawDatCleaned_ctrlfilt = rawDatCleaned_ctrlfilt[! matchID %in% rmv,]

```

Plotting these scaled control values -

```{r, fig.width = 15, fig.height = 5, fig.align="center"}

p = ggplot(ctrlDat) + theme_bw(base_size = 8) + labs(y="Control / Median plate control (roGFP2 ratios) in log2 scale") +
    geom_boxplot(aes(x = Plate, y = NormControl), outlier.size = 0.1, lwd=0.2) +
    facet_wrap( ~ Nutrient) +
    #geom_text_repel(data = subset(ctrlDat, NormControl > 3), aes(x = Plate, y = NormControl, label = Gene), size = 2) +
    theme(axis.text.x = element_text(angle = 90, hjust = 0.5, vjust = 0.5)) 

p

```

A more standard way of calculating the variance among the controls is by computing the robust coefficient of variance [see here](https://stats.stackexchange.com/questions/38635/a-robust-non-parametric-measure-like-coefficient-of-variation-iqr-median-o) and [here](https://arxiv.org/pdf/1907.01110.pdf). For a given plate $i$, this is given as - $$CoefVar_i = \frac{mad(Ctrl_i)}{median(Ctrl_i)}$$ `mad` is `Median absolute deviation`

Plotting these robust coefficient of variation for control values- 

```{r, fig.width = 15, fig.height = 4, fig.align="center"}
ctrlDat$NormControl = 2^ctrlDat$NormControl
cvDat = split(ctrlDat, ctrlDat$Nutrient)
cvDat = lapply(cvDat, function(x){
  a = split(x, x$Plate)
  b = lapply(a, function(y) mad(y$NormControl)/median(y$NormControl))
  return(b)
})

cvDat = melt(cvDat)
cvDat = cvDat[,c(2,3,1)]
colnames(cvDat) = c("Plate", "Nutrient", "Dev")
cvDat$Plate = as.factor(cvDat$Plate)
cvDat$Nutrient = factor(cvDat$Nutrient, levels = c("Glucose", "Galactose", "Glycerol"))

p = ggplot(cvDat, aes(Plate, Dev)) + theme_bw(base_size = 8) + 
labs(y="% of deviation from median plate control roGFP2 ratios") +
geom_bar(stat="identity") + 
facet_wrap( ~ Nutrient) + 
#geom_hline(yintercept = 0.5, color = "black") +
geom_hline(yintercept = 1, color = "black") +
theme(panel.grid = element_blank(), axis.text.x = element_text(angle = 90, hjust = 0.5, vjust = 0.5)) 

ggsave(filename = paste0(path,"analysis/QC/roGFP2_ratio_RAWdata_Controls_Coef_variation.pdf"), plot = p, width = 12, height = 3.5)

p
```

```{r, echo=F}
rm(ctrlDat, cvDat)
```

***

## Saving the data

Finally we save the cleaned raw Data as a `.RDS` (R data object), which will be used for the next step of normalization

```{r, echo=FALSE}

# Raw data to be used for Robust Z normalization
saveRDS(rawDatCleaned, paste0(path,"data/workspaces/YeastRedox_MATa_RawData_ForZNorm.RDS"))
WriteXLS(rawDatCleaned, ExcelFileName = paste0(path,"analysis/supplementary/tables/YeastRedox_MATa_RawData_BackgroundFluorescence_Outliers_Filtered_ToUseForZnorm.xlsx"), 
AdjWidth =TRUE, BoldHeaderRow = TRUE, FreezeRow = 1)

# Raw data to be used for Control based normalization
saveRDS(rawDatCleaned_ctrlfilt, paste0(path,"data/workspaces/YeastRedox_MATa_RawData_ForCtrlNorm.RDS"))
WriteXLS(rawDatCleaned, ExcelFileName = paste0(path,"analysis/supplementary/tables/YeastRedox_MATa_RawData_BackgroundFluorescence_Outliers_PoorControl_Filtered_ToUseForCtrlNorm.xlsx"), 
AdjWidth =TRUE, BoldHeaderRow = TRUE, FreezeRow = 1)

```

## Session information {#SI}

```{r}
sessionInfo()
```
